+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 #### Added
 
 - `unique-by` for more control over how uniqueness is determined.
+- `for` as a better pattern for doing something effectful over the stream.
+
+#### Deprecated
+
+- `for-each`: use `for` instead.
 
 ### 1.3.1 (2025-01-13)